Implement "cached" ratelimits
Remove cache from matchlist
Change if->print to printIf

+++ src/main/java/no/stelar7/api/l4j8/basic/ratelimiting/RateLimit.java
@@ -7,7 +7,7 @@ public class RateLimit
     private int  permits;
     private long delayInMs;
     
-    public RateLimit(final int permits, final int time, final TimeUnit unit)
+    public RateLimit(final int permits, final long time, final TimeUnit unit)
     {
         this.permits = permits;
         delayInMs = unit.toMillis(time);
